patterns/pattern-2/src/classification_function/requirements.txt
PyMuPDF
boto3
pillow